+++ spring-jpa-basic/README.md
@@ -187,3 +187,12 @@
     - toString, lombok, json LIB: 양쪽 toString 메소드에서 순환참조
   - 연관관계의 주인은 외래 키의 위치를 기준으로 정하자
 - 실전 예제 2 - 연관관계 매핑 시작
+
+### 👉 Section 6
+- 다대일 [N:1]
+  - RDB 의 경우 '다' 사이드에 설계상 FK 가 위치한다.
+- 일대다 [1:N]
+  - '일' 이 연관관계의 Owner
+- 일대일 [1:1]
+- 다대다 [N:M]
+- 실전 예제 3 - 다양한 연관관계 매핑
