+++ spring-jpa-basic/src/main/java/hellojpa/Member.java
@@ -1,42 +1,35 @@
-//package hellojpa;
-//
-//import jakarta.persistence.*;
-//
-//@Entity
-//public class Member {
-//
-//    @Id @GeneratedValue
-//    @Column(name = "MEMBER_ID")
-//    private Long id;
-//
-//    @Column(name = "USERNAME")
-//    private String username;
-//
+package hellojpa;
+
+import jakarta.persistence.*;
+
+@Entity
+public class Member {
+
+    @Id @GeneratedValue
+    @Column(name = "MEMBER_ID")
+    private Long id;
+
+    @Column(name = "USERNAME")
+    private String username;
+
 //    @ManyToOne(fetch = FetchType.LAZY)
 //    @JoinColumn(name = "TEAM_ID")
 //    private Team team;
-//
-//    public Team getTeam() {
-//        return team;
-//    }
-//
-//    public void setTeam(Team team) {
-//        this.team = team;
-//    }
-//
-//    public Long getId() {
-//        return id;
-//    }
-//
-//    public void setId(Long id) {
-//        this.id = id;
-//    }
-//
-//    public String getUsername() {
-//        return username;
-//    }
-//
-//    public void setUsername(String username) {
-//        this.username = username;
-//    }
-//}
+
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    public String getUsername() {
+        return username;
+    }
+
+    public void setUsername(String username) {
+        this.username = username;
+    }
+}
