+++ spring-jpa-basic/src/main/java/hellojpa/Team.java
@@ -1,42 +1,46 @@
-//package hellojpa;
-//
-//import jakarta.persistence.*;
-//
-//import java.util.ArrayList;
-//import java.util.List;
-//
-//@Entity
-//public class Team {
-//
-//    @Id @GeneratedValue
-//    @Column(name = "TEAM_ID")
-//    private Long id;
-//    private String name;
-//
+package hellojpa;
+
+import jakarta.persistence.*;
+
+import java.util.ArrayList;
+import java.util.List;
+
+@Entity
+public class Team {
+
+    @Id @GeneratedValue
+    @Column(name = "TEAM_ID")
+    private Long id;
+    private String name;
+
 //    @OneToMany(mappedBy = "team")  // team은 Member의 변수명임
 //    private List<Member> members = new ArrayList<>();
-//
-//    public Long getId() {
-//        return id;
-//    }
-//
-//    public void setId(Long id) {
-//        this.id = id;
-//    }
-//
-//    public String getName() {
-//        return name;
-//    }
-//
-//    public void setName(String name) {
-//        this.name = name;
-//    }
-//
-//    public List<Member> getMembers() {
-//        return members;
-//    }
-//
-//    public void setMembers(List<Member> members) {
-//        this.members = members;
-//    }
-//}
+
+    @OneToMany
+    @JoinColumn(name = "TEAM_ID")
+    private List<Member> members = new ArrayList<>();
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public List<Member> getMembers() {
+        return members;
+    }
+
+    public void setMembers(List<Member> members) {
+        this.members = members;
+    }
+}
