+++ spring-jpa-basic/src/main/java/hellojpa/jpabook/jpashop/JpaMain.java
@@ -1,28 +1,28 @@
-package hellojpa.jpabook.jpashop;
-
-import hellojpa.jpabook.jpashop.domain.Order;
-import jakarta.persistence.EntityManager;
-import jakarta.persistence.EntityManagerFactory;
-import jakarta.persistence.EntityTransaction;
-import jakarta.persistence.Persistence;
-
-public class JpaMain {
-    public static void main(String[] args) {
-        EntityManagerFactory emf = Persistence.createEntityManagerFactory("hello");
-        EntityManager em = emf.createEntityManager();
-
-        EntityTransaction tx = em.getTransaction();
-        tx.begin();
-
-        try {
-            Order order = new Order();
-
-            tx.commit();
-        } catch(Exception e) {
-            tx.rollback();
-        } finally {
-            em.close();
-            emf.close();
-        }
-    }
-}
+//package hellojpa.jpabook.jpashop;
+//
+//import hellojpa.jpabook.jpashop.domain.Order;
+//import jakarta.persistence.EntityManager;
+//import jakarta.persistence.EntityManagerFactory;
+//import jakarta.persistence.EntityTransaction;
+//import jakarta.persistence.Persistence;
+//
+//public class JpaMain {
+//    public static void main(String[] args) {
+//        EntityManagerFactory emf = Persistence.createEntityManagerFactory("hello");
+//        EntityManager em = emf.createEntityManager();
+//
+//        EntityTransaction tx = em.getTransaction();
+//        tx.begin();
+//
+//        try {
+//            Order order = new Order();
+//
+//            tx.commit();
+//        } catch(Exception e) {
+//            tx.rollback();
+//        } finally {
+//            em.close();
+//            emf.close();
+//        }
+//    }
+//}
