+++ spring-jpa-basic/src/main/java/hellojpa/jpabook/jpashop/domain/Member.java
@@ -1,21 +1,21 @@
-package hellojpa.jpabook.jpashop.domain;
-
-import jakarta.persistence.*;
-
-import java.util.ArrayList;
-import java.util.List;
-
-@Entity
-public class Member {
-
-    @Id @GeneratedValue
-    @Column(name = "MEMBER_ID")
-    private Long id;
-    private String name;
-    private String city;
-    private String street;
-    private String zipcode;
-
-    @OneToMany(mappedBy = "member")
-    private List<Order> orders = new ArrayList<>();
-}
+//package hellojpa.jpabook.jpashop.domain;
+//
+//import jakarta.persistence.*;
+//
+//import java.util.ArrayList;
+//import java.util.List;
+//
+//@Entity
+//public class Member {
+//
+//    @Id @GeneratedValue
+//    @Column(name = "MEMBER_ID")
+//    private Long id;
+//    private String name;
+//    private String city;
+//    private String street;
+//    private String zipcode;
+//
+//    @OneToMany(mappedBy = "member")
+//    private List<Order> orders = new ArrayList<>();
+//}
