+++ spring-jpa-basic/src/main/java/hellojpa/jpabook/jpashop/domain/OrderStatus.java
